PytestAPI/requirements.txt
adbutils==0.3.4
airtest==1.1.1
allure-pytest==2.8.6
allure-python-commons==2.8.6
amqp==1.4.9
ansi2html==1.4.2
anyjson==0.3.3
apipkg==1.5
appdirs==1.4.3
APScheduler==3.6.1
asn1crypto==0.24.0
astor==0.8.1
atomicwrites==1.4.0
attrs==19.3.0
concurrent-log-handler==0.9.9
configparser==5.0.0
coverage==4.5.4
cryptography==2.6.1
cryptokit==0.0.7
cssselect==1.0.3
cycler==0.10.0
ddt==1.2.1
decorator==4.4.2
defusedxml==0.5.0
demjson==2.2.4
deprecation==2.0.6
diff-match-patch==20181111
docopt==0.6.2
dwebsocket==0.4.2
elasticsearch==7.8.1
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
eventlet==0.22.1
execnet==1.7.1
fabric==2.5.0
facebook-wda==0.3.4
fake==0.8
fake-useragent==0.1.11
Faker==1.0.2
findit==0.5.7
fire==0.1.3
Flask==1.0.2
flower==0.9.2
future==0.15.2
futures==3.1.1
get==2019.4.13
gevent==1.3.6
greenlet==0.4.15
har2case==0.3.1
hrpc==1.0.8
httpie==1.0.3
httplib2==0.9.2
HttpRunner==1.5.8
humanize==0.5.1
idna==2.9
imageio==2.6.1
importlib-metadata==1.6.1
imutils==0.5.3
invoke==1.3.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jdcal==1.4
jedi==0.15.1
Jinja2==2.10
joblib==0.14.1
jsonpath==0.81
jsonpath-rw==1.4.0
jsonschema==3.0.2
kafka-python==1.4.6
kiwisolver==1.1.0
kombu==3.0.37
locust==0.0
locustio==0.11.0
loguru==0.4.0
logzero==1.5.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.3.0
msgpack==0.5.6
mss==4.0.3
mysql==0.0.2
mysqlclient==1.4.6
namedlist==1.7
nb-log==3.4
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numpy==1.17.0
odfpy==1.4.0
opencv-contrib-python==3.4.2.17
opencv-python==4.1.1.26
openpyxl==2.4.11
packaging==19.1
pandas==0.25.0
pandocfilters==1.4.2
parameterized==0.7.0
paramiko==2.4.1
ParamUnittest==0.2
parse==1.11.1
parso==0.5.1
pickleshare==0.7.5
pika==1.1.0
Pillow==5.2.0
pluggy==0.13.1
ply==3.11
pocoui==1.0.78
post==2019.4.13
prettytable==0.7.2
progress==1.5
progressbar2==3.39.3
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
public==2.0.1
py==1.8.1
pyasn1==0.4.7
pybind11==2.4.3
pycparser==2.19
pycryptodome==3.8.2
pyecharts==1.6.0
pyee==5.0.0
Pygments==2.4.2
pyHook==1.5.1
pymongo==3.5.1
PyMySQL==0.9.3
pyqt5-tools==5.11.3.1.4
pyquery==1.4.0
pyrsistent==0.15.4
pytest==4.5.0
pytest-forked==1.1.3
pytest-html==1.20.0
pytest-metadata==1.8.0
pytest-repeat==0.7.0
pytest-rerunfailures==5.0
pytest-xdist==1.32.0
PyTestReport==0.1.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-utils==2.3.0
PyYAML==5.3.1
records==0.5.3
redis==3.2.1
request==1.0.2
requests==2.23.0
requests-html==0.10.0
requests-toolbelt==0.9.1


